Add chat backend

+++ CampusConnect_Backend/src/main/java/com/campus_connect/CampusConnect_Backend/controllers/NotesController.java
@@ -76,7 +76,7 @@ public class NotesController {
 
 
     @DeleteMapping("/{id}")
-    public void deleteNote(@PathVariable Long id) {
+    public void deleteNote(@PathVariable int id) {
         noteService.deleteNote(id);
     }
