+++ CampusConnect_Backend/src/main/java/com/campus_connect/CampusConnect_Backend/repository/NotesRepository.java
@@ -5,6 +5,6 @@ import com.campus_connect.CampusConnect_Backend.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface NotesRepository extends JpaRepository<Note, Long> {
+public interface NotesRepository extends JpaRepository<Note, Integer> {
     List<Note> findByUploadedBy(User user);
 }
