+++ CampusConnect_Backend/src/main/java/com/campus_connect/CampusConnect_Backend/service/NotesService.java
@@ -22,7 +22,7 @@ public class NotesService {
         return noteRepository.save(note);
     }
 
-    public void deleteNote(Long id) {
+    public void deleteNote(int id) {
         noteRepository.deleteById(id);
     }
